requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.6.0